refactor PostComposer components
