+++ apps/web/src/components/NftSelector/NftSelectorToken.tsx
@@ -3,12 +3,10 @@ import { graphql, useFragment } from 'react-relay';
 import styled from 'styled-components';
 
 import { NftSelectorTokenFragment$key } from '~/generated/NftSelectorTokenFragment.graphql';
-import { useNftRetry } from '~/hooks/useNftRetry';
 import { useTrack } from '~/shared/contexts/AnalyticsContext';
 
 import { NftFailureBoundary } from '../NftFailureFallback/NftFailureBoundary';
-import { NftFailureFallback } from '../NftFailureFallback/NftFailureFallback';
-import { NftSelectorPreviewAsset } from './RawNftSelectorPreviewAsset';
+import { NftSelectorPreviewAsset } from './NftSelectorPreviewAsset';
 
 type Props = {
   tokenRef: NftSelectorTokenFragment$key;
@@ -21,16 +19,13 @@ export function NftSelectorToken({ tokenRef, onSelectToken, isInGroup = false }:
     graphql`
       fragment NftSelectorTokenFragment on Token {
         dbid
-        ...RawNftSelectorPreviewAssetFragment
+        ...NftSelectorPreviewAssetFragment
       }
     `,
     tokenRef
   );
 
   const track = useTrack();
-
-  const { handleNftLoaded, handleNftError, retryKey, refreshMetadata, refreshingMetadata } =
-    useNftRetry({ tokenId: token.dbid });
 
   const handleClick = useCallback(() => {
     if (isInGroup) {
@@ -41,32 +36,14 @@ export function NftSelectorToken({ tokenRef, onSelectToken, isInGroup = false }:
   }, [isInGroup, track, onSelectToken, token]);
 
   return (
-    <NftFailureBoundary
-      key={retryKey}
-      tokenId={token.dbid}
-      fallback={
-        <StyledNftFailureFallbackWrapper>
-          <NftFailureFallback
-            size="medium"
-            tokenId={token.dbid}
-            onRetry={refreshMetadata}
-            refreshing={refreshingMetadata}
-          />
-        </StyledNftFailureFallbackWrapper>
-      }
-      onError={handleNftError}
-    >
-      <NftSelectorPreviewAsset tokenRef={token} onLoad={handleNftLoaded} />
+    <>
       <StyledOutline onClick={handleClick} />
-    </NftFailureBoundary>
+      <NftFailureBoundary tokenId={token.dbid}>
+        <NftSelectorPreviewAsset tokenRef={token} />
+      </NftFailureBoundary>
+    </>
   );
 }
-
-const StyledNftFailureFallbackWrapper = styled.div`
-  height: 100%;
-  width: 100%;
-  position: relative;
-`;
 
 const StyledOutline = styled.div`
   position: absolute;
